requirements (3).txt
# Interface Web
streamlit==1.40.1

# Traitement d'image
opencv-python==4.11.0.86
Pillow==10.4.0
numpy==1.24.3

# Deep Learning
tensorflow==2.13.0

# Suppression d'arrière-plan
rembg==2.0.61

# Analyse faciale
deepface==0.0.93

# Optionnel (pour rembg/ONNX)
onnxruntime==1.19.2

# Gérer et configurer des variables d'environnement
python-dotenv==0.9.9

# Outil de comparaison d'images interactif pour Streamlit
streamlit-image-comparison==0.0.4

# Visualisation (matplotlib)
matplotlib==3.7.5

# Scientific computing (scipy)
scipy==1.10.1